refactor: order history -> user & store
